Refactor team creation logic by moving it from TeamsRepository to UserTeamsRepository. Update related tests to ensure validation error handling for team creation is consistent. Remove obsolete createTeam method from TeamsRepository to streamline functionality.
